ci: clean up codebase
