+++ .github/workflows/e2e-test.yml
@@ -3,10 +3,11 @@ on:
   deployment_status:
 
 jobs:
-  preview:
+  e2eTests:
     if:
       github.event_name == 'deployment_status' &&
       github.event.deployment_status.state == 'success'
+    name: Perform 2e2 tests
     runs-on: ubuntu-latest
     strategy:
       matrix:
