+++ .github/workflows/empty-issues-closer.yaml
@@ -12,6 +12,7 @@ jobs:
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v3 # NOTE: Retrieve issue templates.
+
       - name: Run empty issues closer action
         uses: rickstaa/empty-issues-closer-action@v1
         env:
