+++ .github/workflows/stale-theme-pr-closer.yaml
@@ -5,17 +5,25 @@ on:
 
 jobs:
   closeOldThemePrs:
-    runs-on: ubuntu-latest
     name: Close stale 'invalid' theme PRs
+    runs-on: ubuntu-latest
+    strategy:
+      matrix:
+        node-version: [16.x]
 
     steps:
       - uses: actions/checkout@v3
-      - uses: actions/setup-node@v3
+
+      - name: Setup Node
+        uses: actions/setup-node@v3
         with:
-          node-version: 16
+          node-version: ${{ matrix.node-version }}
+          cache: npm
+
       - uses: bahmutov/npm-install@v1
         with:
           useLockFile: false
+
       - run: npm run close-stale-theme-prs
         env:
           STALE_DAYS: 15
